src/requirements.txt
darksky_weather==1.9.0
delorean==1.0.0
numpy==1.19.5
pandas==1.2.1
requests==2.21.0
